% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APCalignApp-package.R
\docType{package}
\name{APCalignApp-package}
\alias{APCalignApp}
\alias{APCalignApp-package}
\title{APCalignApp: Shiny Interface to R Package APCalign}
\description{
APCalignApp allows you to align and update your Australian Plant Names to currently accepted standards set by the Australian Plant Census
}
\author{
\strong{Maintainer}: Fonti Kar \email{f.kar@unsw.edu.au} (\href{https://orcid.org/0000-0002-2760-3974}{ORCID})

Authors:
\itemize{
  \item Daniel Falster \email{daniel.falster@unsw.edu.au} (\href{https://orcid.org/0000-0002-9814-092X}{ORCID}) [contributor]
  \item Elizabeth Wenk \email{e.wenk@unsw.edu.au} (\href{https://orcid.org/0000-0001-5640-5910}{ORCID}) [contributor]
  \item Will Cornwell \email{w.cornwell@unsw.edu.au} (\href{https://orcid.org/0000-0003-4080-4073}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Carl Boettiger \email{cboettig@gmail.com} (\href{https://orcid.org/0000-0002-1642-628X}{ORCID}) [contributor]
}

}
\keyword{internal}
