% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload.R
\name{upload_file}
\alias{upload_file}
\title{Read uploaded file as a data frame}
\usage{
upload_file(path)
}
\arguments{
\item{path}{path of file}
}
\value{
vector of input names (first column of data frame)
}
\description{
Read uploaded file as a data frame
}
